Disable framebuffer acceleration for debugger
SDL is able to use hardware acceleration for presenting window surface.
This is controlled by SDL_FRAMEBUFFER_ACCELERATION hint, and will use
one of the available graphics API instead of software blitting.

OpenGL context is a state machine. When a texture is bound, it will be
used in all subsequent rendering operations, until a different texture
is bound.

It appears as if somehow both the emulator and debugger windows end up
sharing the context. DOSBox doesn't expect that. As a result, output
gets stalled and corrupted when debugger is invoked. Disabling
framebuffer acceleration feature eliminates that issue.