Setup the MPU-401 regardless of MIDI availability
MPU-401 cards can be present and listening on the ISA
BUS regardless if a MIDI device was plugged in.

if the user wants and MPU-401, then they should get it
even if no MIDI device is present.

(This is actually how DOSBox normally works because
prior to MIDI refactoring, the "none" MIDI device was
still considered a valid and available MIDI device
with respect to the MPU-401 setup routine)