Store and propagate downstream the properties of the video mode
In addition to the currently stored rendered image dimensions and pixel
aspect ratio (PAR), also store the dimensions & PAR of the video mode.
This is necessary for logging the correct "nominal" pixel dimensions and
PAR of a given mode and storing them as metadata in the captured images.

This might also be handy later for certain OSD features that display
information about the current video mode.