Remove reelmagic/vga_passthrough.h
This file would replace subsequent calls and definitions
of certain render functions with reelmagic equivalents.

This means that in files with this included, the render
functions don't call the functions that they say they do.

All calls now call the function directly without this
layer of opaque indirection, and the corresponding
declarations are corrected to match the definitions.