Fix deactivation of queued joystick axis movement
When mapping a host keyboard button to a vitual joystick
axis, the prior logic would allow a queue of the same
event (ie: holding a given direction) to ignore
deactivation of the event (ie: release of the
direction).

This change now zeroes the queued action and deactivates
when requested.

Tested with:
 - Host analog stick driving the axes (native control)
 - Host D-pad mapped to the analog axes
 - Keypad up/down/left/right mapped to analog axes