Remove mechanism to pass #define directives from env vars to shaders
Remove the super obscure "extension mechanism" where #define directives
can be passed to shaders via GLSHADER_* environment variables. This is
originating from the original OpenGL shader SVN patch. Most likely  
nobody uses this for anything and it just complicates the code for no
good reason, so it's best to lose it.