Always write the default config if the primary config file is missing
Prior to this change, if the primary config file was missing, it was
created, but instead of the defaults, the current config settings were
written. This included the local `dosbox.conf` if it existed and any
extra configs loaded with the `--conf` CLI option on top of the actual
defaults.

This change ensures that whenever we create the primary config on
startup, it always contains the defaults and nothing else.