Revert "Add an automatic presentation mode condition for macOS"
This reverts commit 70c108c088f412666e227f1a163fb8b22c32a7ae.

The special casing for macOS causes incorrect output in fullscreen at
120 Hz with the default settings (half-updated frames) when the screen
is only updated sporadically. This is common in older games that update
the screen only a few times per second, or not even update it at all for
as long as there's no user input (which can be several seconds). That's
a sizable chunk of the DOS gaming catalogue, and the macOS special
performance optimisation path broke these games with the default config
in fullscreen @ 120 Hz.

By reverting the macOS special path, we're defaulting to the *correct*
(not the *fastest* possible) behaviour out-of-the-box with everything
left at default/auto for the broadest possible range of games. The whole
auto presentation mode logic also becomes easier to maintain and reason
about.