Restore double scanning support for VGA text modes
This regressed after the major double scanning rework in 766725a5.

Double-scanned VGA text modes are extremely uncommon. A good test case
(and the only one I know of) is the game Indenture (1991) that sets up a
custom 80x25 mode at 640x200 resolution with the 8x8 charset and double
scanning enabled.