Add ESS ES1688 AudioDrive emulation (`sbtype = ess`)
This is an adaptation of the DOSBox-X ESS code originally authored by
Jonathan Campbell.

Only the bare minimum required for the ESS auto-detection routines to
work has been implemented. The ESS ES1688 emulation is supposed to be
used for Enhanced FM Music only; for digital audio, use the Sound
Blaster Pro option in the game's setup.

In general, there's not much to be gained from emulating the ESS DAC.
The ESS cards can operate in Sound Blaster Pro compatibility mode for
digital sound and the sound is identical to that of the SB Pro, except
for perhaps the different sounding low-pass filter (which can be
emulated with custom `sb_filter` settings).

Co-authored-by: John Novak <john@johnnovak.net>

Imported-from:
https://github.com/joncampbell123/dosbox-x/blob/d5743d42a7a891426062d0025e7e24a6991d1df7/src/hardware/sblaster.cpp